optimize recursive/multi-point call exprs
